requirements.txt
flask
requests
openai
PyPDF2
python-pptx
fpdf